+++ src/Transformer/EagerPropertiesResolver/Implementation/DoctrineEagerPropertiesResolver.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Transformer\EagerPropertiesResolver\Implementation;
+
+use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\Mapping\MappingException;
+use Rekalogika\Mapper\Transformer\EagerPropertiesResolver\EagerPropertiesResolverInterface;
+
+class DoctrineEagerPropertiesResolver implements EagerPropertiesResolverInterface
+{
+    public function __construct(private ManagerRegistry $managerRegistry)
+    {
+    }
+
+    public function getEagerProperties(string $sourceClass): array
+    {
+        $manager = $this->managerRegistry->getManagerForClass($sourceClass);
+
+        if (!$manager) {
+            return [];
+        }
+
+        try {
+            $metadata = $manager->getClassMetadata($sourceClass);
+        } catch (\ReflectionException | MappingException) {
+            return [];
+        }
+
+        $identifiers = $metadata->getIdentifierFieldNames();
+
+        return $identifiers;
+    }
+}