+++ tests/Fixtures/Doctrine/EntityWithSingleIdentifier.php
@@ -1,0 +1,89 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\Doctrine;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity]
+class EntityWithSingleIdentifier
+{
+    #[ORM\Id]
+    #[ORM\Column(name: 'my_identifier', type: 'string', length: 255)]
+    private string $myIdentifier;
+
+    #[ORM\Column]
+    private string $name;
+
+    #[ORM\ManyToOne(targetEntity: self::class, inversedBy: 'children')]
+    #[ORM\JoinColumn(name: 'parent_id', referencedColumnName: 'my_identifier')]
+    private ?self $parent = null;
+
+    /**
+     * @var Collection<array-key,self>
+     */
+    #[ORM\OneToMany(targetEntity: self::class, mappedBy: 'parent')]
+    private Collection $children;
+
+    public function __construct(string $myIdentifier, string $name)
+    {
+        $this->myIdentifier = $myIdentifier;
+        $this->name = $name;
+        $this->children = new ArrayCollection();
+    }
+
+    public function getMyIdentifier(): string
+    {
+        return $this->myIdentifier;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function getParent(): ?self
+    {
+        return $this->parent;
+    }
+
+    public function setParent(?self $parent): void
+    {
+        $this->parent = $parent;
+    }
+
+    /**
+     * @return Collection<array-key,self>
+     */
+    public function getChildren(): Collection
+    {
+        return $this->children;
+    }
+
+    public function addChild(self $child): void
+    {
+        if (!$this->children->contains($child)) {
+            $this->children->add($child);
+            $child->setParent($this);
+        }
+    }
+
+    public function removeChild(self $child): void
+    {
+        if ($this->children->removeElement($child)) {
+            $child->setParent(null);
+        }
+    }
+}